ai-lead-scoring-pipeline/requirements.txt
pandas
numpy
python-dotenv
gspread
oauth2client
openai
biopython